backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
apscheduler==3.10.4
playwright==1.42.0
beautifulsoup4==4.12.3
httpx==0.27.0
